Use standard init.rc
BoardConfig.mk

For now we are using the old skool surfaceflinger style. This is setup
with the following build environment variables in the BoardConfig.mk

	COMMON_GLOBAL_CFLAGS += -DEGL_NEEDS_FNW
	TARGET_RUNNING_WITHOUT_SYNC_FRAMEWORK := true
	TARGET_DISABLE_TRIPLE_BUFFERING := true
	TARGET_DOESNT_USE_FENCE_SYNC := true

Support needs to be added to the gralloc/hwcomposer to use the sync
framework. Fencing Sync requires functionality to be added to the EGL
implementation, no sure about triple buffering.

H/W vsync would need to be added to the kernel framebuffer driver